+++ cartridges/int_klaviyo/cartridge/templates/default/klaviyo/klaviyoFooter.isml
@@ -15,9 +15,6 @@
         window.klaviyoSFCCSiteID = "${klaviyoUtils.siteId}";
         window.klaviyoSFCCArchitecture = 'site-genesis';
     </script>
-    <isif condition="${klaviyoUtils.customerHubEnabled}">
-        <isinclude template="customerHub/customerHubBootstrap" />
-    </isif>
     <isif condition="${pageContext.type == 'product' ||
         pageContext.type == 'search' ||
         pageContext.type == 'storefront'
