--- cartridges/int_klaviyo_customer_hub/cartridge/int_klaviyo_customer_hub.properties
@@ -1,2 +1,0 @@
-## cartridge.properties for cartridge int_klaviyo_customer_hub
-demandware.cartridges.int_klaviyo_customer_hub.id=int_klaviyo_customer_hub
