--- cartridges/int_klaviyo_customer_hub/cartridge/templates/default/customerHub/customerHubBootstrap.isml
@@ -1,55 +1,0 @@
-<isscript>
-    var URLUtils = require('dw/web/URLUtils');
-    var klaviyoUtils = require('*/cartridge/scripts/klaviyo/utils');
-    var activeProductBuilder = require('*/cartridge/scripts/customerHub/activeProduct');
-
-    var bootstrapData = null;
-
-    if (klaviyoUtils.customerHubEnabled) {
-        var productId = '';
-
-        if (typeof pageContext !== 'undefined' && pageContext.type === 'product') {
-            if (typeof pdict !== 'undefined' && pdict.Product && pdict.Product.ID) {
-                productId = pdict.Product.ID;
-            }
-        }
-
-        bootstrapData = {
-            storefrontRoutes: {
-                login: URLUtils.url('Login-Show').toString(),
-                register: URLUtils.url('Login-Show').toString(),
-                profile: URLUtils.url('Account-Show').toString(),
-                addresses: URLUtils.url('Address-List').toString()
-            }
-        };
-
-        if (productId) {
-            var activeProduct = activeProductBuilder.buildActiveProduct(String(productId).trim());
-            if (activeProduct) {
-                bootstrapData.activeProduct = activeProduct;
-            }
-        }
-    }
-
-    var bootstrapJson = null;
-
-    if (bootstrapData) {
-        bootstrapJson = JSON.stringify(bootstrapData)
-            .replace(/</g, '\\u003c')
-            .replace(/\u2028/g, '\\u2028')
-            .replace(/\u2029/g, '\\u2029');
-    }
-</isscript>
-<isif condition="${bootstrapJson}">
-    <script>
-        if (!window.customerHub) {
-            window.customerHub = {};
-        }
-        (function (bootstrapData) {
-            window.customerHub.storefrontRoutes = bootstrapData.storefrontRoutes;
-            if (bootstrapData.activeProduct) {
-                window.customerHub.activeProduct = bootstrapData.activeProduct;
-            }
-        })(<isprint value="${bootstrapJson}" encoding="off" />);
-    </script>
-</isif>
