--- cartridges/int_klaviyo_customer_hub_core/cartridge/int_klaviyo_customer_hub_core.properties
@@ -1,2 +1,0 @@
-## cartridge.properties for cartridge int_klaviyo_customer_hub_core
-demandware.cartridges.int_klaviyo_customer_hub_core.id=int_klaviyo_customer_hub_core
